Create archinstall.sh
This is a script that I have written for Arch Linux, with your BSPWM customisations. This is the first script that I've ever made, so it's definitely not exactly perfect. But I have done my best to make it so that it's at the very least usable, and I even managed to include a couple of things in the script so that users can have a little more options out of the box. Just to make things a little easier. For example, I have included the bibata cursor theme, Breeze cursors, Oxygen, and Nordzy. I also have an XFCE theme included called Greybird, that can work with the GUI based apps, such as Nautilus, Variety, and such. Speaking of which, I also have Nautilus and Variety for file management and for changing the wallpaper.